+++ api/models/__init__.py
@@ -1,0 +1,8 @@
+from .patient_model import Patient
+from .visit_model import Visit
+from .vitals_modal import Vitals
+from .consult_model import Consult
+from .diagnosis_model import Diagnosis
+from .medication_model import Medication
+from .order_model import Order
+from django.contrib.auth.models import User
