+++ api/tests/test_consult.py
@@ -43,6 +43,7 @@ class TestConsultAPI(TestSetup):
                     "status": "status",
                 },
                 "doctor": {"username": "test_user"},
+                "prescriptions": [],
                 "date": "2021-01-01T00:00:00Z",
                 "past_medical_history": "past_medical_history",
                 "consultation": "consultation",
@@ -81,6 +82,7 @@ class TestConsultAPI(TestSetup):
                     "status": "status",
                 },
                 "doctor": {"username": "test_user"},
+                "prescriptions": [],
                 "date": "2021-01-01T00:00:00Z",
                 "past_medical_history": "past_medical_history",
                 "consultation": "consultation",
@@ -122,6 +124,7 @@ class TestConsultAPI(TestSetup):
                     "status": "status",
                 },
                 "doctor": {"username": "test_user"},
+                "prescriptions": [],
                 "date": "2021-01-01T00:00:00Z",
                 "past_medical_history": "past_medical_history",
                 "consultation": "consultation",
@@ -160,6 +163,7 @@ class TestConsultAPI(TestSetup):
                     "status": "status",
                 },
                 "doctor": {"username": "test_user"},
+                "prescriptions": [],
                 "date": "2021-01-01T00:00:00Z",
                 "past_medical_history": "past_medical_history",
                 "consultation": "consultation",
