+++ api/tests/test_orders.py
@@ -36,37 +36,7 @@ class TestOrdersAPI(TestSetup):
                 "quantity": 1,
                 "notes": "notes",
             },
-            "consult": {
-                "id": 1,
-                "visit": {
-                    "id": 1,
-                    "patient": {
-                        "model": "clinicmodels.patient",
-                        "pk": 1,
-                        "village_prefix": "VPF",
-                        "name": "patient_name",
-                        "identification_number": "identification_number",
-                        "contact_no": "contact_no",
-                        "gender": "gender",
-                        "date_of_birth": "2021-01-01T00:00:00Z",
-                        "drug_allergy": "drug_allergy",
-                        "face_encodings": None,
-                        "picture": "image/upload/v1715063294/ghynewr4gdhkuttombwc.jpg",
-                        "filter_string": "VPF001VPF1 contact_no patient_name",
-                        "patient_id": "VPF001",
-                    },
-                    "date": "2021-01-01T00:00:00Z",
-                    "status": "status",
-                },
-                "doctor": {"username": "test_user"},
-                "date": "2021-01-01T00:00:00Z",
-                "past_medical_history": "past_medical_history",
-                "consultation": "consultation",
-                "plan": "plan",
-                "referred_for": "referred_for",
-                "referral_notes": "referral_notes",
-                "remarks": "remarks",
-            },
+            "consult": 1,
             "quantity": 100,
             "notes": "order_notes",
             "remarks": None,
